Refactoring marker overlay
